refactored inference stuff
